requirements.txt
pystray
pillow
pynput
PyAudio
soundfile
ollama
pyperclip
pyautogui
requests
sv-ttk
cohere
openai
playsound==1.2.2
kokoro-onnx
psutil
cryptography
